refactor utils, inputs and constants into /common
