+++ app/metro.config.js
@@ -1,4 +1,11 @@
 const {getDefaultConfig, mergeConfig} = require('@react-native/metro-config');
+const path = require('path');
+const extraNodeModules = {
+  'common': path.resolve(__dirname + '/../common'),
+};
+const watchFolders = [
+  path.resolve(__dirname + '/../common')
+];
 
 /**
  * Metro configuration
@@ -6,6 +13,11 @@ const {getDefaultConfig, mergeConfig} = require('@react-native/metro-config');
  *
  * @type {import('metro-config').MetroConfig}
  */
-const config = {};
+const config = {
+  resolver: {
+    extraNodeModules,
+  },
+  watchFolders,
+};
 
 module.exports = mergeConfig(getDefaultConfig(__dirname), config);
