+++ app/package.json
@@ -39,6 +39,7 @@
     "@types/react-native-dotenv": "^0.2.0",
     "@types/react-test-renderer": "^18.0.0",
     "babel-jest": "^29.2.1",
+    "babel-plugin-module-resolver": "^5.0.0",
     "eslint": "^8.19.0",
     "jest": "^29.2.1",
     "metro-react-native-babel-preset": "0.76.7",
