--- app/types/passportData.ts
@@ -1,28 +1,0 @@
-export type MrzInfo = {
-  compositeCheckDigit: string;
-  dateOfBirth: string;
-  dateOfBirthCheckDigit: string;
-  dateOfExpiry: string;
-  dateOfExpiryCheckDigit: string;
-  documentCode: string;
-  documentNumber: string;
-  documentNumberCheckDigit: string;
-  documentType: number;
-  gender: string;
-  issuingState: string;
-  nationality: string;
-  optionalData1: string;
-  primaryIdentifier: string;
-  secondaryIdentifier: string;
-};
-
-export type DataHash = [number, number[]];
-
-export type PassportData = {
-  mrz: string;
-  signatureAlgorithm: string;
-  pubKey: {modulus?: string, curveName?: string, publicKeyQ?: string};
-  dataGroupHashes: DataHash[];
-  eContent: number[];
-  encryptedDigest: number[];
-};
