+++ common/.gitignore
@@ -1,0 +1,2 @@
+inputs
+node_modules/
