+++ common/README.md
@@ -1,0 +1,3 @@
+# Proof of Passport Commons
+
+Constants and utils used in multiple Proof of Passport subdirectories.
