+++ common/package.json
@@ -1,0 +1,10 @@
+{
+  "dependencies": {
+    "@babel/runtime": "^7.23.4",
+    "js-sha256": "^0.10.1",
+    "node-forge": "^1.3.1"
+  },
+  "devDependencies": {
+    "@types/node-forge": "^1.3.10"
+  }
+}
