+++ common/src/constants/constants.ts
@@ -1,0 +1,9 @@
+export const attributeToPosition = {
+  issuing_state: [2, 4],
+  name: [5, 43],
+  passport_number: [44, 52],
+  nationality: [54, 56],
+  date_of_birth: [57, 62],
+  gender: [64, 64],
+  expiry_date: [65, 70],
+}
