+++ common/src/utils/types.ts
@@ -1,0 +1,10 @@
+export type DataHash = [number, number[]];
+
+export type PassportData = {
+  mrz: string;
+  signatureAlgorithm: string;
+  pubKey: {modulus?: string, curveName?: string, publicKeyQ?: string};
+  dataGroupHashes: DataHash[];
+  eContent: number[];
+  encryptedDigest: number[];
+};
