+++ common/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "resolveJsonModule": true,
+    "esModuleInterop": true,
+    "target": "ES2015",
+    "moduleResolution": "node"
+  }
+}
